chore: Move agent function to agent router
