+++ src/main.py
@@ -17,6 +17,7 @@ from src.response.handler import exception_handler
 from auth.router import router as auth_router
 from project.router import router as project_router
 from user.router import router as user_router
+from agent.router import router as agent_router
 
 
 app = FastAPI(
@@ -61,18 +62,8 @@ async def get_documentation(
 
 
 
-@app.get("/generate_issues")
-async def generate_issues():
-    """
-    Generate issues using the agent executor.
-    """
-
-    executor = chain.CustomAgentExecutor()
-    result = await executor.generate_issues()
-    return result
-
-
 # Include routers
 app.include_router(auth_router)
 app.include_router(project_router)
 app.include_router(user_router)
+app.include_router(agent_router)
